Adding missing functions for other document types
