+++ lib/businesscards/getBusinessCard.js
@@ -1,18 +1,18 @@
 const Client = require('../client/Client');
 /**
- * Get a w9 document https://docs.veryfi.com/api/w9s/get-a-w-9/
+ * Get a specific Business card. https://docs.veryfi.com/api/business-cards/get-a-business-card/
  * @memberof Client
  * @param {number} document_id The unique identifier of the document.
  * @param {boolean} bounding_boxes A field used to determine whether to return bounding_box and bounding_region for extracted fields in the Document response.
  * @param {boolean} confidence_details A field used to determine whether to return the score and ocr_score fields in the Document response.
  * @returns {JSON} Object of a previously processed blueprinted document.
  */
-Client.prototype.get_w9 = async function (
+Client.prototype.get_business_card = async function (
     document_id,
     bounding_boxes = false,
     confidence_details = false
 ) {
-    let endpoint_name = `/w9s/${document_id}/`;
+    let endpoint_name = `/business-cards/${document_id}/`;
     let request_arguments = {
         "bounding_boxes": bounding_boxes,
         "confidence_details": confidence_details
