+++ lib/checks/getChecks.js
@@ -1,19 +1,23 @@
 const Client = require('../client/Client');
 /**
- * Get a w9 document https://docs.veryfi.com/api/w9s/get-a-w-9/
+ * Get all checks. https://docs.veryfi.com/api/checks/get-checks/
  * @memberof Client
- * @param {number} document_id The unique identifier of the document.
+ * @param {number} page The page number. The response is capped to maximum of 50 results per page.
+ * @param {number} page_size The number of Documents per page.
  * @param {boolean} bounding_boxes A field used to determine whether to return bounding_box and bounding_region for extracted fields in the Document response.
  * @param {boolean} confidence_details A field used to determine whether to return the score and ocr_score fields in the Document response.
- * @returns {JSON} Object of a previously processed blueprinted document.
+ * @returns {JSON} Object of previously processed any documents
  */
-Client.prototype.get_w9 = async function (
-    document_id,
+Client.prototype.get_checks = async function (
+    page = 1,
+    page_size = 50,
     bounding_boxes = false,
     confidence_details = false
 ) {
-    let endpoint_name = `/w9s/${document_id}/`;
+    let endpoint_name = `/checks/`;
     let request_arguments = {
+        "page": page,
+        "page_size": page_size,
         "bounding_boxes": bounding_boxes,
         "confidence_details": confidence_details
     };
