+++ lib/checks/processCheckBuffer.js
@@ -1,23 +1,27 @@
 const Client = require('../client/Client');
 /**
- * Process a w9 document from an url. https://docs.veryfi.com/api/w9s/process-a-w-9/
+ * Process Check and extract all the fields from it. https://docs.veryfi.com/api/bank-statements/process-a-bank-statement/
+ *
  * @memberof Client
- * @param {String} file_url url file to submit for data extraction
+ * @param {String} file_name The file name including the extension
+ * @param {String} file_buffer Buffer of a file to submit for data extraction
  * @param {boolean} bounding_boxes A field used to determine whether to return bounding_box and bounding_region for extracted fields in the Document response.
  * @param {boolean} confidence_details A field used to determine whether to return the score and ocr_score fields in the Document response.
  * @param {Object} kwargs Additional request parameters
  * @returns {JSON} Data extracted from the document
  */
-Client.prototype.process_w9_from_url = async function (
-    file_url,
+Client.prototype.process_check_from_buffer = async function (
+    file_name,
+    file_buffer,
     bounding_boxes = false,
     confidence_details = false,
     {...kwargs} = {}
 ) {
 
-    let endpoint_name = "/w9s/";
+    let endpoint_name = "/checks/";
     let request_arguments = {
-        "file_url": file_url,
+        "file_name": file_name,
+        "file_data": file_buffer,
         "bounding_boxes": bounding_boxes,
         "confidence_details": confidence_details,
     };
