+++ lib/types/BoundingElement.ts
@@ -1,0 +1,6 @@
+export declare type BoundingElement = {
+    bounding_box?: null | number[];
+    bounding_region?: null | number[];
+    rotation?: null | number;
+    value?: null | number | string;
+}
