+++ lib/types/Client.ts
@@ -1,178 +1,7 @@
 import fs from "fs";
-import path from "path";
-
-/**
- * Additional request parameters type
- */
-export declare type VeryfiExtraArgs = Record<
-    string | number | symbol,
-    string | number | boolean
->;
-
-export declare type BoundingElement = {
-  bounding_box?: null | number[];
-  bounding_region?: null | number[];
-  rotation?: null | number;
-  value?: null | number | string;
-}
-
-export type JsonObject = Record<string, any>;
-
-/**
- * Object of data extracted from the document
- */
-export declare type VeryfiDocument = {
-  abn_number?: null | string | BoundingElement;
-  account_number?: null | string | BoundingElement;
-  accounting_entry_type?: null | string;
-  balance?: null | string | BoundingElement;
-  barcodes?: null | string[] | BoundingElement[];
-  bill_to?: null | BillTo;
-  card_number?: null | string | BoundingElement;
-  cashback?: null | number | BoundingElement;
-  category?: null | string | BoundingElement;
-  created_date?: null | string;
-  currency_code?: null | string | BoundingElement;
-  custom_fields?: null | VeryfiExtraArgs;
-  date?: null | string | BoundingElement;
-  default_category?: null | string | BoundingElement;
-  delivery_date?: null | string | BoundingElement;
-  delivery_note_number?: null | string | BoundingElement;
-  discount?: null | number | BoundingElement;
-  document_reference_number?: null | string | BoundingElement;
-  document_title?: null | string | BoundingElement;
-  document_type?: null | string | BoundingElement;
-  due_date?: null | string | BoundingElement;
-  duplicate_of?: null | number;
-  exch_rate?: null | number;
-  external_id?: null | string;
-  final_balance?: null | number | BoundingElement;
-  guest_count?: null | number | BoundingElement;
-  id?: null | number;
-  img_blur?: null | boolean;
-  img_file_name?: null | string;
-  img_thumbnail_url?: null | string;
-  img_url?: null | string;
-  incoterms?: null | number | BoundingElement;
-  insurance?: null | number | BoundingElement;
-  invoice_number?: null | string | BoundingElement;
-  is_approved?: null | boolean;
-  is_blurry?: null | boolean[];
-  is_document?: null | boolean;
-  is_duplicate?: null | boolean;
-  is_money_in?: null | boolean | BoundingElement;
-  license_plate_number?: null | string | BoundingElement;
-  line_items?: null | LineItem[];
-  line_items_with_scores?: null | LineItem[];
-  notes?: null | string;
-  ocr_text?: null | string;
-  order_date?: null | string | BoundingElement;
-  payment?: null | Payment;
-  pdf_url?: null | string;
-  previous_balance?: null | number | BoundingElement;
-  purchase_order_number?: null | string | BoundingElement;
-  rounding?: null | number | BoundingElement;
-  server_name?: null | string | BoundingElement;
-  service_end_date?: null | string;
-  service_start_date?: null | string;
-  ship_date?: null | string | BoundingElement;
-  ship_to?: null | ShipTo;
-  shipping?: null | number | BoundingElement;
-  status?: null | string;
-  store_number?: null | string | BoundingElement;
-  subtotal?: null | number | BoundingElement;
-  tags?: null | Tag[];
-  tax?: null | number | BoundingElement;
-  tax_lines?: null | TaxLine[];
-  tip?: null | number | BoundingElement;
-  total?: null | number | BoundingElement;
-  total_pages?: null | number;
-  total_quantity?: null | string | BoundingElement;
-  total_weight?: null | string | BoundingElement;
-  tracking_number?: null | string | BoundingElement;
-  tracking_numbers?: null | BoundingElement[];
-  updated_date?: null | string;
-  vending_person?: null | string | BoundingElement;
-  vending_person_number?: null | string | BoundingElement;
-  vat_number?: null | string;
-  vendor?: null | Vendor;
-  vendors?: null | BoundingElement[];
-  vin_number?: null | string | BoundingElement;
-  warnings?: null | string[];
-  weights?: null | BoundingElement[];
-};
-
-declare type LineItem = {
-  date?: null | string | BoundingElement;
-  description?: null | string | BoundingElement;
-  discount?: null | number | BoundingElement;
-  end_date?: null | string | BoundingElement;
-  hsn?: null | string;
-  id?: null | number;
-  order?: null | number;
-  price?: null | number | BoundingElement;
-  quantity?: null | number | BoundingElement;
-  reference?: null | string | BoundingElement;
-  section?: null | string;
-  sku?: null | string | BoundingElement;
-  start_date?: null | string;
-  tags?: null | null[];
-  tax?: null | number | BoundingElement;
-  tax_rate?: null | number;
-  total?: null | number | BoundingElement;
-  type?: null | string;
-  unit_of_measure?: null | string;
-};
-
-declare type TaxLine = {
-  base?: null | number | BoundingElement;
-  name?: null | string | BoundingElement;
-  order?: null | number;
-  rate?: null | number | BoundingElement;
-  total?: null | number | BoundingElement;
-};
-
-declare type Vendor = {
-  address?: null | string | BoundingElement;
-  category?: null | string | BoundingElement;
-  email?: null | string | BoundingElement;
-  fax_number?: null | string | BoundingElement;
-  name?: null | string | BoundingElement;
-  phone_number?: null | string | BoundingElement;
-  raw_name?: null | string;
-  vendor_logo?: null | string | BoundingElement;
-  vendor_reg_number?: null | string | BoundingElement;
-  vendor_type?: null | string;
-  web?: null | string | BoundingElement;
-};
-
-declare type Payment = {
-  card_number?: null | string | BoundingElement;
-  display_name?: null | string | BoundingElement;
-  terms?: null | number | BoundingElement;
-  type?: null | string;
-};
-
-declare type Tag = {
-  id?: null | number;
-  name?: null | string;
-};
-
-declare type BillTo = {
-  address?: null | string | BoundingElement;
-  email?: null | string | BoundingElement;
-  name?: null | string | BoundingElement;
-  parsed_address?: null | string;
-  phone_number?: null | number | string | BoundingElement;
-  reg_number?: null | number | string | BoundingElement;
-  vat_number?: null | number | string | BoundingElement;
-};
-
-declare type ShipTo = {
-  address?: null | string | BoundingElement;
-  name?: null | string | BoundingElement;
-  parsed_address?: null | string;
-};
+import {VeryfiDocument} from "./VeryfiDocument";
+import {JsonObject, VeryfiExtraArgs} from "./VeryfiExtraArgs";
+import {Tag} from "./Tag";
 
 export declare class Client {
   /**
@@ -195,27 +24,16 @@ export declare class Client {
       api_version?: string,
       timeout?: number
   );
-
   client_id: string;
-
   client_secret: string | undefined;
-
   username: string;
-
   api_key: string;
-
   base_url: string;
-
   api_version: string;
-
   CATEGORIES: string[];
-
   private _get_headers;
-
   private _get_url;
-
   private _generate_signature;
-
   private _request;
 
   /**
@@ -234,9 +52,10 @@ export declare class Client {
    * Retrieve document by ID
    * @memberof Client
    * @param {string} document_id ID of the document you'd like to retrieve
+   * @param {Object} kwargs Additional request parameters
    * @returns {Promise<VeryfiDocument>} Object of data extracted from the document
    */
-  public get_document(document_id: string): Promise<VeryfiDocument>;
+  public get_document(document_id: string,  {...kwargs}?: VeryfiExtraArgs): Promise<VeryfiDocument>;
 
   /**
    * Process a document and extract all the fields from it
@@ -360,40 +179,14 @@ export declare class Client {
       {...kwargs}?: VeryfiExtraArgs
   ): Promise<VeryfiDocument>;
 
-  /**
-   * Add a new tag on an existing document
-   *
-   * @param {number} document_id ID of the document you'd like to add a Tag
-   * @param {string} tag name to add
-   * @return {Promise<Tag>} response about tag added.
-   */
-  public add_tag(document_id: string, tag: string): Promise<Tag>;
 
   /**
-   * Delete all tags on an existing document
-   *
-   * @param {number} document_id ID of the document you'd like to delete all Tags
-   * @return {Promise<any>} response about deleted tags.
+   * Delete any document from Veryfi
+   * @memberof Client
+   * @param {string} document_id ID of the document you'd like to delete
+   * @returns {Promise<VeryfiDocument>} Object of data extracted from the document
    */
-  public delete_tags(document_id: string): Promise<any>;
-
-  /**
-   * Add multiple tags on an existing document
-   *
-   * @param {number} document_id ID of the document you'd like to add a Tag
-   * @param {string[]} tags name to add
-   * @return {Promise<Tag>} response about tags added.
-   */
-  public add_tags(document_id: string, tags: string[]): Promise<Tag>;
-
-  /**
-   * Replace multiple tags on an existing document
-   *
-   * @param {number} document_id ID of the document you'd like to add a Tag
-   * @param {string[]} tags names to be added
-   * @return {Promise<Tag>} response about tags added.
-   */
-  public replace_tags(document_id: string, tags: string[]): Promise<Tag>;
+  public delete_any_document(document_id: string): Promise<any>;
 
   /**
    * Process any document and extract all the fields from it
@@ -446,9 +239,15 @@ export declare class Client {
    * Get a specific any document
    * @memberof Client
    * @param {number} document_id The unique identifier of the document.
+   * @param {boolean} bounding_boxes A field used to determine whether to return bounding_box and bounding_region for extracted fields in the Document response.
+   * @param {boolean} confidence_details A field used to determine whether to return the score and ocr_score fields in the Document response.
    * @returns {Promise<JsonObject>} Object of a previously processed blueprinted document.
    */
-  public get_any_document(document_id: number): Promise<JsonObject[]>;
+  public get_any_document(document_id: number,
+                          bounding_boxes?: boolean,
+                          confidence_details?: boolean): Promise<JsonObject[]>;
+
+
 
   /**
    * Get a specific bank statement
@@ -514,6 +313,42 @@ export declare class Client {
       confidence_details?: boolean,
       {...kwargs}?: VeryfiExtraArgs
   ): Promise<JsonObject>;
+
+
+  /**
+   * Add a new tag on an existing document
+   *
+   * @param {number} document_id ID of the document you'd like to add a Tag
+   * @param {string} tag name to add
+   * @return {Promise<Tag>} response about tag added.
+   */
+  public add_tag(document_id: string, tag: string): Promise<Tag>;
+
+  /**
+   * Delete all tags on an existing document
+   *
+   * @param {number} document_id ID of the document you'd like to delete all Tags
+   * @return {Promise<any>} response about deleted tags.
+   */
+  public delete_tags(document_id: string): Promise<any>;
+
+  /**
+   * Add multiple tags on an existing document
+   *
+   * @param {number} document_id ID of the document you'd like to add a Tag
+   * @param {string[]} tags name to add
+   * @return {Promise<Tag>} response about tags added.
+   */
+  public add_tags(document_id: string, tags: string[]): Promise<Tag>;
+
+  /**
+   * Replace multiple tags on an existing document
+   *
+   * @param {number} document_id ID of the document you'd like to add a Tag
+   * @param {string[]} tags names to be added
+   * @return {Promise<Tag>} response about tags added.
+   */
+  public replace_tags(document_id: string, tags: string[]): Promise<Tag>;
 
 }
