+++ lib/types/Payment.ts
@@ -1,0 +1,8 @@
+import {BoundingElement} from "./BoundingElement";
+
+export declare type Payment = {
+    card_number?: null | string | BoundingElement;
+    display_name?: null | string | BoundingElement;
+    terms?: null | number | BoundingElement;
+    type?: null | string;
+};
