+++ lib/types/ShipTo.ts
@@ -1,0 +1,7 @@
+import {BoundingElement} from "./BoundingElement";
+
+export declare type ShipTo = {
+    address?: null | string | BoundingElement;
+    name?: null | string | BoundingElement;
+    parsed_address?: null | string;
+};
