+++ lib/types/Tag.ts
@@ -1,0 +1,4 @@
+export declare type Tag = {
+    id?: null | number;
+    name?: null | string;
+};
