+++ lib/types/Vendor.ts
@@ -1,0 +1,15 @@
+import {BoundingElement} from "./BoundingElement";
+
+export declare type Vendor = {
+    address?: null | string | BoundingElement;
+    category?: null | string | BoundingElement;
+    email?: null | string | BoundingElement;
+    fax_number?: null | string | BoundingElement;
+    name?: null | string | BoundingElement;
+    phone_number?: null | string | BoundingElement;
+    raw_name?: null | string;
+    vendor_logo?: null | string | BoundingElement;
+    vendor_reg_number?: null | string | BoundingElement;
+    vendor_type?: null | string;
+    web?: null | string | BoundingElement;
+};
