+++ lib/types/VeryfiExtraArgs.ts
@@ -1,0 +1,9 @@
+/**
+ * Additional request parameters type
+ */
+export declare type VeryfiExtraArgs = Record<
+    string | number | symbol,
+    string | number | boolean
+>;
+
+export type JsonObject = Record<string, any>;
