+++ lib/w2s/deleteW2.js
@@ -5,7 +5,7 @@ const Client = require('../client/Client');
  * @memberof Client
  * @param {string} document_id ID of the document you'd like to delete
  */
-Client.prototype.delete_w2_document = async function (document_id) {
+Client.prototype.delete_w2 = async function (document_id) {
     let endpoint_name = `/w2s/${document_id}/`;
     let request_arguments = {"id": document_id};
     return this._request("DELETE", endpoint_name, request_arguments);
