+++ lib/w2s/getW2.js
@@ -5,7 +5,7 @@ const Client = require('../client/Client');
  * @param {string} document_id ID of the document you'd like to retrieve
  * @returns {JSON} Data extracted from the Document
  */
-Client.prototype.get_w2_document = async function (
+Client.prototype.get_w2 = async function (
     document_id,
 ) {
     let endpoint_name = `/w2s/${document_id}/`;
