+++ lib/w2s/getW2s.js
@@ -5,7 +5,7 @@ const Client = require('../client/Client');
  * @param {int} page number, response is capped to maximum of 50 results per page.
  * @return {Array} An array of JSON with all w2 documents.
  */
-Client.prototype.get_w2_documents = async function (page = null) {
+Client.prototype.get_w2s = async function (page = null) {
     let endpoint_name = "/w2s/";
     let request_arguments = {};
     let params = {};
