+++ lib/w2s/processW2.js
@@ -9,7 +9,7 @@ const fs = require('fs');
  * @param {Object} kwargs Additional request parameters
  * @return {JSON} Data extracted from the document.
  */
-Client.prototype.process_w2_document = async function (
+Client.prototype.process_w2 = async function (
     file_path,
     delete_after_processing = false,
     max_pages_to_process = 1,
