+++ lib/w2s/processW2fromBuffer.js
@@ -9,7 +9,7 @@ const Client = require('../client/Client');
  * @param {Object} kwargs Additional request parameters
  * @return {JSON} Data extracted from the document.
  */
-Client.prototype.process_w2_document_from_buffer = async function (
+Client.prototype.process_w2_from_buffer = async function (
     file_name,
     file_buffer,
     delete_after_processing = null,
