+++ lib/w2s/processW2fromUrl.js
@@ -11,7 +11,7 @@ const Client = require('../client/Client');
  * @param {Object} kwargs Additional request parameters
  * @return {JSON} Data extracted from the document.
  */
-Client.prototype.process_w2_document_from_url = async function (
+Client.prototype.process_w2_from_url = async function (
     file_name,
     file_url,
     file_urls = null,
