+++ lib/w8bene/processW8BENEUrl.js
@@ -1,6 +1,9 @@
 const Client = require('../client/Client');
 /**
- * Process a w9 document from an url. https://docs.veryfi.com/api/w9s/process-a-w-9/
+ * Process W-8BEN-E and extract all the fields from it. https://docs.veryfi.com/api/w-8ben-e/process-a-w-8-ben-e/
+ * @example
+ * veryfi_client.process_w8bene_url('file_url')
+ *
  * @memberof Client
  * @param {String} file_url url file to submit for data extraction
  * @param {boolean} bounding_boxes A field used to determine whether to return bounding_box and bounding_region for extracted fields in the Document response.
@@ -8,14 +11,14 @@ const Client = require('../client/Client');
  * @param {Object} kwargs Additional request parameters
  * @returns {JSON} Data extracted from the document
  */
-Client.prototype.process_w9_from_url = async function (
+Client.prototype.process_w8bene_url = async function (
     file_url,
     bounding_boxes = false,
     confidence_details = false,
     {...kwargs} = {}
 ) {
 
-    let endpoint_name = "/w9s/";
+    let endpoint_name = "/w-8ben-e/";
     let request_arguments = {
         "file_url": file_url,
         "bounding_boxes": bounding_boxes,
