+++ lib/w9s/getW9s.js
@@ -1,17 +1,26 @@
 const Client = require('../client/Client');
 /**
  * Get all w9 documents. https://docs.veryfi.com/api/w9s/get-w-9-s/
- * @memberOf Client
- * @param {int} page number, response is capped to maximum of 50 results per page.
- * @return {Array} An array of JSON with all w2 documents.
+ * @memberof Client
+ * @param {number} page The page number. The response is capped to maximum of 50 results per page.
+ * @param {number} page_size The number of Documents per page.
+ * @param {boolean} bounding_boxes A field used to determine whether to return bounding_box and bounding_region for extracted fields in the Document response.
+ * @param {boolean} confidence_details A field used to determine whether to return the score and ocr_score fields in the Document response.
+ * @returns {JSON} Object of previously processed any documents
  */
-Client.prototype.get_w9_documents = async function (page = null) {
-    let endpoint_name = "/w9s/";
-    let request_arguments = {};
-    let params = {};
-    if (page !== null) {
-        params = {"page": page};
-    }
-    let response = await this._request("GET", endpoint_name, request_arguments, params);
-    return response['data']['results'];
+Client.prototype.get_w9s = async function (
+    page = 1,
+    page_size = 50,
+    bounding_boxes = false,
+    confidence_details = false
+) {
+    let endpoint_name = `/w9s/`;
+    let request_arguments = {
+        "page": page,
+        "page_size": page_size,
+        "bounding_boxes": bounding_boxes,
+        "confidence_details": confidence_details
+    };
+    let document = await this._request("GET", endpoint_name, request_arguments);
+    return document['data'];
 }
